Split the regridding step from the calculation
